Refactor parallelization (#19)
* Parallel upload of files/packages (instead of splitting a single file into smaller parts);
* Ensure sending signatures after packages;
* Get the building sequence from the routines' lists;
* Count build threads properly;
* External (systemd's unit) frozen notifier;
* Say remaining packages count in the frozen notification.

New time limits (systemd only):
* hourly: 4hrs
* morning/afternoon/nightly: 6hrs
* tkg-kernels: 24hrs
* tkg-wine: 6hrs

+++ src/lib/base-make.sh
@@ -5,7 +5,7 @@ function lowerstrap() {
 
   local _LOCK_FN _LOCK_FD
 
-  install -o"$(whoami)" -dDm755 "$CAUR_LOWER_DIR"
+  [[ ! -e "$CAUR_LOWER_DIR" ]] && install -o"$(whoami)" -dDm755 "$CAUR_LOWER_DIR"
 
   _LOCK_FN="${CAUR_LOWER_DIR}/lock"
   touch "${_LOCK_FN}"
